pyproject.toml
[tool.poetry]
name = "erroneous-payment-detection"
version = "0.1.0"
description = "An intelligent system for detecting erroneous payments using ML, RAG, and AI agents"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
# ML & Data Science
numpy = "^1.26.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
torch = "^2.1.0"
transformers = "^4.35.0"
xgboost = "^2.0.0"
lightgbm = "^4.1.0"

# RAG & Vector Store (in-memory)
langchain = "^0.1.0"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.0"

# MCP & Agents
anthropic = "^0.8.0"
openai = "^1.6.0"
pydantic = "^2.5.0"
httpx = "^0.25.0"

# API & Backend
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
redis = "^5.0.0"

# Data Processing
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
structlog = "^24.1.0"

# Excel Support
openpyxl = "^3.1.0"
xlrd = "^2.0.1"
pyodbc = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
pre-commit = "^3.5.0"

[tool.poetry.group.ui.dependencies]
streamlit = "^1.28.0"
plotly = "^5.18.0"
dash = "^2.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
